// src/types/anamnesis.ts
export type AnamnesisFormData = {
  lifeAnamnesis: string
  attackAnamnesis: string
  allergyAnamnesis: string
  complains: string
  medicationsTaken: string
  generalState: string
  behavior: string
  condition: string
  anisocoria: string
  nystagmus: string
  skinState: string
  breathing: string
  wheezes: string
  heartSounds: string
  pulse: string
  tongue: string
  belly: string
  additionalObjectiveData: string
  neurologicalStatus: string
}